requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
django-cors-headers==4.6.0
Pillow==10.4.0
psycopg[binary]==3.2.3
gunicorn==22.0.0
whitenoise==6.7.0
dj-database-url==2.2.0
python-dotenv==1.0.1
